# docs/openapi-spec.yaml
openapi: 3.0.0
info:
  title: Tekana eWallet APIs
  contact:
    name: Benit Havugimana
    email: jajabenit@gmail.com
  version: 1.0.0
servers:
- url: https://virtserver.swaggerhub.com/JAJABENIT/Tekana/1.0.0
  description: SwaggerHub API Auto Mocking
- url: http://localhost:3000
tags:
- name: customer
- name: wallet
- name: transaction
paths:
  /customer/register:
    post:
      tags:
      - customer
      summary: Register
      operationId: customerRegisterPOST
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                fullName: jaja
                nationalId: "1234556667"
                gender: M
                email: jajabenit@gmail.com
                password: Password@10
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}
      x-swagger-router-controller: Customer
  /customer/login:
    put:
      tags:
      - customer
      summary: Login
      operationId: customerLoginPUT
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                email: jajabenit@gmail.com
                password: Password@10
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}
      x-swagger-router-controller: Customer
  /customer/{customerId}:
    get:
      tags:
      - customer
      summary: Customer by Id
      operationId: customerCustomerIdGET
      parameters:
      - name: customerId
        in: path
        description: Numeric ID of the user to get
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}
      security:
      - bearerAuth: []
      x-swagger-router-controller: Customer
  /customer:
    get:
      tags:
      - customer
      summary: Customers
      operationId: customerGET
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}
      security:
      - bearerAuth: []
      x-swagger-router-controller: Customer
  /wallet:
    get:
      tags:
      - wallet
      summary: Accounts By Customer
      operationId: walletGET
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}
      security:
      - bearerAuth: []
      x-swagger-router-controller: Wallet
    post:
      tags:
      - wallet
      summary: Create Account
      operationId: walletPOST
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                walletType: personal
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}
      security:
      - bearerAuth: []
      x-swagger-router-controller: Wallet
  /wallet/deposit:
    post:
      tags:
      - wallet
      summary: Add Balance for Testing
      operationId: walletDepositPOST
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                accNumber: 200-6-1672993748555
                amount: 10000
                transactionId: 290002
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}
      security:
      - bearerAuth: []
      x-swagger-router-controller: Wallet
  /wallet/{accNumber}:
    get:
      tags:
      - wallet
      summary: Account By Acc Number
      operationId: walletAccNumberGET
      parameters:
      - name: accNumber
        in: path
        description: String Acc Number of the wallet to get
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}
      security:
      - bearerAuth: []
      x-swagger-router-controller: Wallet
  /transaction:
    post:
      tags:
      - transaction
      summary: Create Transaction
      operationId: transactionPOST
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                receiverAcc: 200-6-1672993748555
                senderAcc: 200-6-1672999556649
                amount: 10
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}
      security:
      - bearerAuth: []
      x-swagger-router-controller: Transaction
  /transaction/{accNumber}:
    get:
      tags:
      - transaction
      summary: Transactions By Acc Number
      operationId: transactionAccNumberGET
      parameters:
      - name: accNumber
        in: path
        description: String Acc Number of the wallet to get
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}
      security:
      - bearerAuth: []
      x-swagger-router-controller: Transaction
components:
  schemas: {}
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer